Remove legacy multiplex tests aggregator
