+++ crates/protocol/src/multiplex/tests/send/vectored/mod.rs
@@ -1,0 +1,4 @@
+use super::*;
+
+mod errors;
+mod paths;
