+++ docs/resume_note.md
@@ -1,0 +1,15 @@
+# Resume Notes
+
+## Status
+- Work paused before implementation changes could begin due to token quota exhaustion during previous session.
+- Repository instructions remain authoritative; no code modules were modified in this pass.
+
+## Suggested Next Steps
+1. Re-read `AGENTS.md` at the repository root to ensure compliance with global conventions, especially branding (oc-rsync/oc-rsyncd) and centralized messaging.
+2. Audit current binaries and config paths to ensure `oc-` prefixes and `/etc/oc-rsyncd/oc-rsyncd.conf` locations are consistently applied.
+3. Verify CI workflow aligns with single Linux test job plus parallel cross-compilation targets as required by the consolidated mission brief.
+4. Continue production release readiness tasks: close remaining protocol/feature parity gaps, enforce documentation standards, and reconfirm coverage ≥95%.
+
+## Notes
+- No tests were executed in this session.
+- Pending tasks should reference `docs/production_scope_p1.md` to maintain convergence discipline.
